Using debug logger; Had to refactor some errors into ./src/errors/test-run/shared.js
