+++ package.json
@@ -54,8 +54,6 @@
     "prepublishOnly": "publish-please guard"
   },
   "dependencies": {
-    "@types/diff": "^4.0.2",
-    "@types/node": "^10.12.19",
     "async-exit-hook": "^1.1.2",
     "babel-core": "^6.22.1",
     "babel-plugin-transform-class-properties": "^6.24.1",
@@ -141,10 +139,12 @@
     "@types/chai": "^3.5.2",
     "@types/debug": "^4.1.5",
     "@types/dedent": "^0.7.0",
+    "@types/diff": "^4.0.2",
     "@types/json5": "0.0.30",
     "@types/mime-db": "^1.27.0",
     "@types/mustache": "^0.8.32",
     "@types/nanoid": "^2.1.0",
+    "@types/node": "^10.12.19",
     "@types/source-map-support": "^0.5.0",
     "@types/useragent": "^2.1.1",
     "@typescript-eslint/eslint-plugin": "1.13.0",
