+++ src/client/driver/command-executors/execute-child-window-selector.js
@@ -1,5 +1,5 @@
 import hammerhead from '../deps/hammerhead';
-import { ChildWindowNotFoundError } from '../../../errors/test-run';
+import { ChildWindowNotFoundError } from '../../../errors/test-run/shared';
 
 const Promise = hammerhead.Promise;
