+++ src/client/driver/driver-link/iframe/parent.js
@@ -3,7 +3,7 @@ import {
     EstablishConnectionMessage,
     CommandExecutedMessage
 } from '../messages';
-import { CurrentIframeIsNotLoadedError } from '../../../../errors/test-run';
+import { CurrentIframeIsNotLoadedError } from '../../../../errors/test-run/shared';
 import sendMessageToDriver from '../send-message-to-driver';
 import { WAIT_FOR_IFRAME_DRIVER_RESPONSE_TIMEOUT } from '../timeouts';
 import sendConfirmationMessage from '../send-confirmation-message';
