+++ src/client/driver/driver-link/window/child.js
@@ -1,6 +1,6 @@
 import { CloseAllChildWindowsMessage, SetAsMasterMessage } from '../messages';
 import sendMessageToDriver from '../send-message-to-driver';
-import { CannotSwitchToWindowError, CloseChildWindowError } from '../../../../errors/test-run';
+import { CannotSwitchToWindowError, CloseChildWindowError } from '../../../../errors/test-run/shared';
 import { WAIT_FOR_WINDOW_DRIVER_RESPONSE_TIMEOUT } from '../timeouts';
 
 export default class ChildWindowDriverLink {
