--- src/utils/diff/const.ts
@@ -1,1 +1,0 @@
-export const INDENT = '   ';
